refactor: 팩토리 구조 재조직 - 도메인별 fixture 폴더로 분리
1. TestEntityFactoriesIntegrationTest 이동:
   - bottlenote-product-api/src/test/.../fixture/
     → bottlenote-product-api/src/test/.../operation/

2. bottlenote-mono 팩토리 파일들을 각 도메인 폴더로 이동:
   - UserTestFactory → user/fixture/
   - AlcoholTestFactory → alcohols/fixture/
   - RatingTestFactory → rating/fixture/
   - ReviewTestFactory → review/fixture/
   - BusinessSupportTestFactory → support/business/fixture/
   - LikesTestFactory → like/fixture/
   - PicksTestFactory → picks/fixture/
   - HistoryTestFactory → history/fixture/

3. 각 팩토리 파일의 package 선언 수정

4. TestEntityFactoriesIntegrationTest의 import 경로 수정

5. Rule.md → TestFactoryRule.md로 이름 변경 및 상위 폴더로 이동

6. 빈 fixture 폴더 제거
