+++ bottlenote-mono/src/test/java/app/bottlenote/TestFactoryRule.md
