+++ bottlenote-mono/src/test/java/app/bottlenote/rating/fixture/RatingTestFactory.java
@@ -1,4 +1,4 @@
-package app.bottlenote.fixture;
+package app.bottlenote.rating.fixture;
 
 import app.bottlenote.alcohols.domain.Alcohol;
 import app.bottlenote.rating.domain.Rating;
